Change all line endings to LF.
OpenVMS supports the CRLF ending in the original file, but the VMS
git client only handles Stream_LF files.

While it's annoying to replace every line of a file in a commit
in order to change the line endings, I may as well do it now and
get it out of the way.

I also removed some trailing spaces and spaces intermixed with tabs
that git diff highlighted for me in red.